+++ js/pages/login/auth.js
@@ -28,6 +28,6 @@ export function clearAuth() {
   localStorage.removeItem(ACCESS_KEY);
   localStorage.removeItem(REFRESH_KEY);
   localStorage.removeItem("user");
+  sessionStorage.removeItem("order_type");
   sessionStorage.removeItem("orderList");
-  sessionStorage.removeItem("orderType");
 }
